refactor(auth): move Permission/Role/ROLE_PERMISSIONS to autobot_shared (#6511)
Eliminates the security drift risk where the Permission enum and
ROLE_PERMISSIONS dict lived only in autobot-backend/auth_rbac.py with no
mechanism to keep the SLM backend in sync.

Changes:
- NEW autobot_shared/auth/permissions.py — canonical Permission enum (43
  members), Role enum (6 members), ROLE_PERMISSIONS dict; single source of
  truth for both backends
- autobot_shared/auth/__init__.py — re-exports Permission/Role/ROLE_PERMISSIONS
  alongside existing JWT helpers
- autobot-backend/auth_rbac.py — replaces inline class definitions with
  `from autobot_shared.auth.permissions import ...`; backward-compat
  re-exports maintained so all 65 Depends(get_current_user) callers are
  unaffected
- NEW autobot-backend/tests/security/test_cross_service_auth_parity.py — 15
  tests enforcing enum/role-mapping invariants; any new permission that is
  not granted to at least one role causes a test failure (prevents silent
  drift)

Follow-up sub-issues filed:
- #7568 — SLM in-process permission cache → Redis (Phase 2)
- #7569 — auth architecture documentation (Phase 3)
